requirements.txt
guizero
pycryptodome